requirements.txt
tensorflow-cpu
streamlit
librosa
gdown
matplotlib
pillow
